Kennelly Power Model

Calculates running power based on distance using Kennelly's model and di Prampero's equation.

Kennelly_1906_P(
  Distance,
  Ck,
  C1 = 3.86,
  C2 = 0.4,
  C3 = 2,
  BMR = 1.2,
  BSA = 1.8,
  BM = 70,
  eff = 0.2
)

Arguments

Distance

Numeric. Running distance in meters (must be positive).

Ck

Numeric. Empirical coefficient determined from a set of performances (must be positive).

C1

Numeric. Parameter of the equation. Default is 3.86.

C2

Numeric. Parameter of the equation. Default is 0.4.

C3

Numeric. Parameter of the equation. Default is 2.

BMR

Numeric. Basal metabolic rate in W/kg. Default is 1.2.

BSA

Numeric. Body surface area in m^2. Default is 1.8.

BM

Numeric. Body mass in kg. Default is 70.

eff

Numeric. Assumed mechanical efficiency. Default is 0.2.

Value

Numeric. Power output in watts.

References

Kennelly, A. E. "An Approximate Law of Fatigue in the Speeds of Racing Animals." Proceedings of the American Academy of Arts and Sciences 42, no. 15 (1906): 275–331. https://doi.org/10.2307/20022230.

Examples

Kennelly_1906_P(Distance = 1000, Ck = 10)
#> [1] 257.583