Critical Power Model

This function calculates the power output based on the Critical Power model.

critical_power_model(t, cp, Wprime)

Arguments

t

Numeric. Running duration in seconds.

cp

Numeric. Critical power in w.

Wprime

Numeric. Anaerobic energy reserve in J.

Value

Numeric. Power output in watts.

References

Hill, D. W. "The Critical Power Concept. A Review." Sports Medicine 16, no. 4 (October 1993): 237–54. https://doi.org/10.2165/00007256-199316040-00003.

Monod, H., and J. Scherrer. "The Work Capacity of a Synergic Muscle Group." Ergonomics 8, no. 3 (July 1, 1965): 329–38. https://doi.org/10.1080/00140136508930810.

Moritani, T., A. Nagata, H. A. deVries, and M. Muro. "Critical Power as a Measure of Physical Work Capacity and Anaerobic Threshold." Ergonomics 24, no. 5 (May 1981): 339–50. https://doi.org/10.1080/00140138108924856.

Examples

critical_power_model(t = 300, cp = 300, Wprime = 20000)
#> [1] 366.6667