Roy and Joyner Running Speed Model

This function calculates world record running speed based on the Roy and Joyner model.

model_roy(S, V0, B, b, d, Distance)

Arguments

S

Numeric. Index representing the aerobic metabolic slowdown or reduced contribution as running distance increases (m/s). Must be positive.

V0

Numeric. Extrapolated speed at 100 m (m/s). Must be positive.

B

Numeric. Index representing an anaerobic boost at 100 m (m/s). Must be positive.

b

Numeric. Determines the shape of the anaerobic boost curve as a function of distance. Must be positive.

d

Numeric. Rate of decline of the anaerobic boost with increasing distance (m). Must be positive.

Distance

Numeric. Running distance in meters. Must be positive.

Value

Numeric. World record running speed in m/s.

References

Roy, Tuhin K., Michael J. Joyner, Jonathon W. Senefeld, Chad C. Wiggins, and Timothy W. Secomb. "An Empirical Model for World Record Running Speeds with Distance, Age, and Sex: Anaerobic and Aerobic Contributions to Performance." Journal of Applied Physiology (Bethesda, Md.: 1985), June 27, 2024. https://doi.org/10.1152/japplphysiol.00033.2024.

Examples

model_roy(S = 0.1, V0 = 12, B = 2, b = 1.5, d = 200, Distance = 1000)
#> [1] 11.99199